👌 Rearrange code
