+++ upgrades/0.0.103.eks-1-21/pkg/kubectl/version/testdata/kubectlVersion.yaml
