+++ upgrades/0.0.103.eks-1-21/pkg/kubectl/version/types.go
@@ -1,0 +1,6 @@
+package version
+
+type Versions struct {
+	ClientVersion Version `yaml:"clientVersion"`
+	ServerVersion Version `yaml:"serverVersion"`
+}
